LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.object.update-expected.txt

2d.gradient.object.update


FAIL OffscreenCanvas test: 2d.gradient.object.update assert_approx_equals: Red channel of the pixel at (50, 25) expected 0 +/- 2 but got 255
